Updates after Udemy course section 67-71 - various
