+++ app.env
@@ -1,5 +1,4 @@
 ENVIRONMENT=development
-DB_DRIVER=pgx
 DB_SOURCE=postgresql://admin:adminSecret@localhost:5432/simple_bank?sslmode=disable
 MIGRATION_URL=file://db/migration
 HTTP_SERVER_ADDRESS=0.0.0.0:8080
