+++ db/sqlc/error.go
@@ -7,6 +7,14 @@ import (
 	"github.com/jackc/pgx/v5/pgconn"
 )
 
+/*
+  Find these in the postgres documentation
+
+  Even though we are using the pgx, there could be some
+  pg code which might take too long to remove.
+
+*/
+
 const (
 	ForeignKeyViolation = "23503"
 	UniqueViolation     = "23505"
