+++ db/sqlc/store.go
@@ -3,7 +3,6 @@ package db
 import (
 	"context"
 
-	"github.com/jackc/pgx/v5"
 	"github.com/jackc/pgx/v5/pgxpool"
 )
 
@@ -40,23 +39,4 @@ func NewStore(pool *pgxpool.Pool) Store {
 		Queries: New(pool),
 		db:      pool,
 	}
-}
-
-func (store *SQLStore) execTx(ctx context.Context, fn func(*Queries) error) error {
-	tx, err := store.db.BeginTx(ctx, pgx.TxOptions{})
-	if err != nil {
-		return err
-	}
-
-	q := New(tx)
-
-	err = fn(q)
-	if err != nil {
-		if rbErr := tx.Rollback(ctx); rbErr != nil {
-			return rbErr
-		}
-		return err
-	}
-
-	return tx.Commit(ctx)
 }
